# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bike Trip Journal API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/trips/{slug}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get a trip's metadata and its bikes",
        "description": "**Context.** A trip from before accounts was shared through two links, a rider\nlink and a viewer link. Since decision-log Entry 29 either link only *locates*\nthe trip; what the caller may do is decided by their membership. This is the\nendpoint both links open first. It answers three questions in one round trip \u2014\nwhich trip is this, which bikes are on it, and should the UI offer writes \u2014\nbecause the app is used on the Stuart Hwy where a second request is a second\nchance to be offline. Tasks `t-trip-metadata-endpoint`, `t-am-write-gate-legacy`.\n\n**How it works.** `{slug}` is resolved by the shared `require_trip_access`\ndependency, which accepts **either** slug and raises a 404 if neither matches.\nThe session is optional here and never produces a 401. The dependency derives\n`access`: `rider` iff the session user has an active membership on this trip\n(read fresh on every request), `viewer` for everyone else \u2014 anonymous, a\nsigned-in non-member, a pending requester or a revoked member. Which slug was\nfollowed plays no part. `viewer.role` is derived alongside it (`anonymous`,\n`none`, `pending`, `rider` or `leader`), and `access` is `rider` exactly when\nthat role is `rider` or `leader`. `visibility`, `publicDelayHours`, `riderCount`\nand `lastPublicStopAt` are filled as on `GET /api/v2/trips/{tripId}`; this read\nitself stays full and undelayed for either slug. The trip's bikes are fetched by `data/repositories/bikes.list_by_trip`,\nfiltered to this trip; a trip with no bikes returns `\"bikes\": []`, which is a\nnormal trip and not a 404. Neither slug is in the response: `TripOut` has no\nslug field, and the slug is the credential.\n\n`access` is a **UI hint**, deprecated, and not the enforcement point. It tells\nthe frontend whether to render Add stop / upload photo / edit bikes. The write\nendpoints check the session and membership themselves regardless of what the UI\nchose to show.\n\n**Related APIs.** `GET /api/trips/{slug}/stops` for the stops this header sits\nabove, `GET /api/trips/{slug}/map` for the same trip as GeoJSON, and\n`POST /api/trips/{slug}/bikes` / `PATCH /api/trips/{slug}/bikes/{id}` for the\nmember-only writes behind the `bikes` list returned here.",
        "operationId": "get_trip_api_trips__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip.",
              "title": "Slug"
            },
            "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip."
          }
        ],
        "responses": {
          "200": {
            "description": "The trip, the bikes on it, and whether the caller is an active member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this slug. Deliberately the same answer for a mistyped link, a revoked one and a guess \u2014 see `docs/api-contract.md`, 'Access control: 403 and 404 are different answers'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "A path parameter failed validation. This operation takes no request body, so the path is the only thing that can fail; its path parameters are plain strings, so in practice it does not return this today.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{slug}/stops": {
      "get": {
        "tags": [
          "stops"
        ],
        "summary": "List a trip's stops",
        "description": "**Context.** The stops are the journal \u2014 where the rider got to, when, and what\nthey wrote about it. `GET /api/trips/{slug}` is the header above this list;\nthis is the body. It is also the **only** endpoint that serves `notes` and\n`locationSource`: `GET /api/trips/{slug}/map` deliberately omits both so the\nsame data is not served from two places and given two chances to drift, and the\nfrontend resolves a clicked pin back to its full stop from this list by `id`.\nTask `t-stops-list-endpoint`.\n\n**How it works.** `{slug}` is resolved by the shared `require_trip_access`\ndependency, which accepts **either** slug and raises a 404 if neither matches; a\nviewer slug is an ordinary success here, not a 403, because this is a read. The\nstops are fetched by `data/repositories/stops.list_by_trip`, filtered on the\n`trip_id` the dependency resolved rather than on anything the caller typed, so\nno other trip's stops can appear. **A trip with no stops is a `200` with `[]`** \u2014\nan empty collection, not a missing one, and never a 404. Neither slug is in the\nresponse: `StopOut` has no slug field, and the slug is the credential.\n\nThe list comes back ordered by `arrivedAt` then `id`, but that is for response\nstability only and is **not** part of the contract \u2014 match stops by `id`, never\nby position. The map endpoint's chronological trail is established by the map\nhandler itself and does not read its ordering guarantee from here.\n\n**Related APIs.** `GET /api/trips/{slug}` for the trip header and its bikes,\n`POST /api/trips/{slug}/stops` to add a stop (active members only),\n`GET /api/trips/{slug}/stops/{id}/photos` for one stop's photos, and\n`GET /api/trips/{slug}/map` for the same stops as GeoJSON.",
        "operationId": "list_stops_api_trips__slug__stops_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip.",
              "title": "Slug"
            },
            "description": "The trip's rider or viewer slug \u2014 the link the trip was shared with. Either one locates the trip; neither grants anything. Legacy reads accept both and give the full, undelayed trip."
          }
        ],
        "responses": {
          "200": {
            "description": "Every stop on this trip, with its notes and how it was located.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StopOut"
                  },
                  "title": "Response List Stops Api Trips  Slug  Stops Get"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this slug. Deliberately the same answer for a mistyped link, a revoked one and a guess \u2014 see `docs/api-contract.md`, 'Access control: 403 and 404 are different answers'. A trip that exists but has no stops yet is **not** this case: that is a `200` with `[]`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "A path parameter failed validation. This operation takes no request body, so the path is the only thing that can fail; its path parameters are plain strings, so in practice it does not return this today.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `public-read` limit: 120 requests a minute per client address. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stops"
        ],
        "summary": "Add a stop to a trip",
        "description": "**Context.** This is how a stop gets into the journal, and the one write the app\nmakes most: the rider taps \"Add stop\" on the Stuart Hwy, often with no signal, so\nthe stop is captured on the device with an id the *client* generates and queued\nuntil there is a connection. Only a signed-in, active member of the trip can add\none; either slug just locates the trip (decision-log Entry 29, contract default\n21). Tasks `t-stops-create-endpoint`, `t-am-write-gate-legacy`.\n\n**How it works.** `require_trip_writer` checks, in order: the slug (404 if no\ntrip has it), the session (401), and an active membership on the trip, read fresh\non every request (403 \"You're not a rider on this trip.\" or, for a revoked\nmember, \"You're no longer a rider on this trip.\"). The stop is stored against the\n`trip_id` that dependency resolved, never against anything in the path or the\nbody, with `created_by` set to the signed-in account. `StopCreate.id` then decides\none of **three** branches\n(`docs/api-contract.md`, \"Idempotency\"; decision-log Entry 14):\n\n- an **unseen** id creates the stop \u2014 `201`, body is the new stop;\n- an id **already on this trip** is a replay \u2014 `200`, body is the **stored**\n  record, and a body that differs from the stored one changes nothing;\n- an id that exists **on a different trip** is not a replay \u2014 `409` /\n  `CONFLICT`, nothing is created, and nothing about the other record is\n  disclosed.\n\nThe replay is matched on `(trip_id, id)` rather than on the id alone, so no\nother trip's stop can be returned through this trip's slug, and the third branch\nis decided by a check *before* the insert rather than by letting the insert fail\n\u2014 a driver error would render `500` / `INTERNAL_ERROR`, which the offline queue\nretries forever.\n\n`arrivedAt` is the time the rider arrived, not the time this request was\nreceived, and its **UTC offset is required**: a naive value is a `422`, because a\nrider crossing timezones has no offset worth guessing (decision-log Entry 15).\n\n**Related APIs.** `GET /api/trips/{slug}/stops` lists what this endpoint writes,\n`GET /api/trips/{slug}` is the trip header above it (its `access` field is the\nUI hint for whether to offer this write at all: `rider` iff the caller is an\nactive member),\n`POST /api/trips/{slug}/stops/{id}/photos` attaches photos to a stop created\nhere, and `GET /api/trips/{slug}/map` renders these stops as GeoJSON.",
        "operationId": "create_stop_api_trips__slug__stops_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's rider or viewer slug. It only locates the trip: the write itself needs a signed-in account with an active membership on it. An unknown slug is a 404, no session a 401, and a non-member or revoked member a 403.",
              "title": "Slug"
            },
            "description": "The trip's rider or viewer slug. It only locates the trip: the write itself needs a signed-in account with an active membership on it. An unknown slug is a 404, no session a 401, and a non-member or revoked member a 403."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stop as stored \u2014 the one just created (201), or the one already stored under this id (200).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopOut"
                }
              }
            }
          },
          "200": {
            "description": "**A replay, not a second stop.** This id is already a stop on this trip, so nothing was created and the **stored** record is returned unchanged \u2014 even where this request's body differs from it. A success, not an error: the offline queue can resend a create it never saw confirmed, any number of times, and reconcile against whichever attempt landed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopOut"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. Checked after the slug and before membership. If a session cookie was sent it is cleared (`Max-Age=0`). The offline queue pauses on this and resumes after sign-in. Nothing was written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, and the slug located the trip, but the account has no active membership on it: \"You're not a rider on this trip.\" for a non-member, \"You're no longer a rider on this trip.\" for a revoked one. The slug grants nothing, whichever one it is. Nothing was written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this slug. Deliberately the same answer for a mistyped link, a revoked one and a guess, and checked before the session \u2014 see `docs/api-contract.md`, 'Access control: 401, 403 and 404 are three different answers'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The `id` in the body already belongs to a stop on a **different** trip, so this is not a replay. Nothing was created, and nothing about the conflicting record is disclosed \u2014 not in the body, not in the message. Never-retry for the offline queue: the same id will conflict on every future attempt, so the stop needs a new one.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The body failed schema validation \u2014 a missing or mistyped field, a `locationSource` outside `gps`/`manual`, or an `arrivedAt` with **no UTC offset**. That last one is server-enforced only; the generated client types it as a plain string and cannot catch it. Also an `id` containing a NUL character, which no stored id can hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `writes` limit: 600 requests an hour per account, or per client address for a request with no session. Checked before the slug, the session and the membership, so nothing was written. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trips/{slug}/bikes": {
      "post": {
        "tags": [
          "bikes"
        ],
        "summary": "Add a bike to a trip",
        "description": "**Context.** Bikes are registered per trip so the journal records who is riding\nwhat. Only a signed-in, active member of the trip can add one; either slug just\nlocates the trip (decision-log Entry 29, contract default 21). The server sets\n`created_by` from the session. Tasks `t-bikes-create-endpoint`,\n`t-am-write-gate-legacy`.\n\n**How it works.** Same three-way idempotency branch as `POST /trips/{slug}/stops`\n(decision-log Entry 14): unseen id creates the bike (201), id already on this\ntrip is a replay (200, stored record returned unchanged), id on a different trip\nis a 409 with nothing disclosed about the conflicting record. `require_trip_writer`\nruns first: slug (404), session (401), active membership (403).\n\n**Related APIs.** `GET /api/trips/{slug}` returns bikes in `TripOut.bikes`,\n`PATCH /api/trips/{slug}/bikes/{id}` edits a bike created here.",
        "operationId": "create_bike_api_trips__slug__bikes_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The trip's rider or viewer slug. It only locates the trip: the write itself needs a signed-in account with an active membership on it. An unknown slug is a 404, no session a 401, and a non-member or revoked member a 403.",
              "title": "Slug"
            },
            "description": "The trip's rider or viewer slug. It only locates the trip: the write itself needs a signed-in account with an active membership on it. An unknown slug is a 404, no session a 401, and a non-member or revoked member a 403."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BikeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The bike as stored -- the one just created (201), or the one already stored under this id (200).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeOut"
                }
              }
            }
          },
          "200": {
            "description": "**A replay, not a second bike.** This id is already a bike on this trip, so nothing was created and the **stored** record is returned unchanged. Task `t-bikes-create-endpoint`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BikeOut"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: none sent, expired, signed out or revoked, or the account is disabled. Checked after the slug and before membership. If a session cookie was sent it is cleared (`Max-Age=0`). The offline queue pauses on this and resumes after sign-in. Nothing was written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, and the slug located the trip, but the account has no active membership on it: \"You're not a rider on this trip.\" for a non-member, \"You're no longer a rider on this trip.\" for a revoked one. The slug grants nothing, whichever one it is. Nothing was written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No trip has this slug.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The `id` in the body already belongs to a bike on a **different** trip. Nothing was created, nothing about the conflicting record is disclosed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The body failed schema validation, or its `id` contains a NUL character, which no stored id can hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "The `writes` limit: 600 requests an hour per account, or per client address for a request with no session. Checked before the slug, the session and the membership, so nothing was written. Retry after `Retry-After` seconds.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait before retrying; always at least 1.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Access": {
        "type": "string",
        "enum": [
          "rider",
          "viewer"
        ],
        "title": "Access"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "TripOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The trip's id. Stable, and safe to show or log \u2014 it is not a credential and cannot be used to reach the trip. Neither slug appears anywhere in this response."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The trip's display name, as the rider titled it."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate",
            "description": "The day the trip starts, as an ISO 8601 date (YYYY-MM-DD). A date with no time and no timezone: it is the calendar day the journal opens on, not an instant, so it reads the same wherever the viewer is."
          },
          "bikes": {
            "items": {
              "$ref": "#/components/schemas/BikeOut"
            },
            "type": "array",
            "title": "Bikes",
            "description": "The bikes on **this** trip, embedded so the app's first load is one request rather than two. Empty list when no bikes have been added yet \u2014 that is a normal trip, not an error. Order is not part of the contract; match bikes by `id`."
          },
          "access": {
            "$ref": "#/components/schemas/Access",
            "description": "Deprecated; kept for clients from before accounts. 'rider' when the caller is signed in with an active membership on this trip (rider or leader), 'viewer' otherwise \u2014 anonymous, a signed-in non-member, a pending requester or a revoked member. Which slug was used plays no part. A UI hint for whether to show write UI (Add stop, upload photo, edit bikes); the server enforces access on every write regardless."
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "description": "Who can read the trip: 'public' (anyone, with the public delay applied for non-members) or 'private' (members only; anyone else gets the same 404 as for a trip that doesn't exist)."
          },
          "publicDelayHours": {
            "type": "integer",
            "title": "Publicdelayhours",
            "description": "How many hours a stop stays hidden from non-members after its `arrivedAt`, from 0 to 168. Members see every stop at once, whatever this says."
          },
          "riderCount": {
            "type": "integer",
            "title": "Ridercount",
            "description": "How many people are on the trip: its active members, leaders included. A count only; who they are is never in this response."
          },
          "lastPublicStopAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastpublicstopat",
            "description": "The latest `arrivedAt` among the stops visible to the public, i.e. with the public delay applied, as a timezone-aware ISO 8601 instant. The same value for every caller, members included. Null when no stop is visible to the public yet."
          },
          "viewer": {
            "$ref": "#/components/schemas/ViewerOut",
            "description": "What the caller is to this trip, computed from the session on every request. Replaces `access`. A UI hint only: the server enforces access on every write regardless."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "startDate",
          "bikes",
          "access",
          "visibility",
          "publicDelayHours",
          "riderCount",
          "lastPublicStopAt",
          "viewer"
        ],
        "title": "TripOut",
        "description": "One trip: `GET /api/trips/{slug}` (either slug) and `GET /api/v2/trips/{tripId}`.\n\nCarries no slug, no username and no user id -- a non-member can receive it."
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "public",
          "private"
        ],
        "title": "Visibility",
        "description": "Who can read a trip. Mirrors `trips_visibility_check` in migration 0003."
      }
    }
  }
}